requirements.txt
fastapi
uvicorn
python-multipart
bcrypt
google-cloud-firestore
google-cloud-datastore
google-cloud-storage
google-auth
google-cloud-aiplatform
requests
streamlit